requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
Markdown==2.6.5
djangorestframework==3.3.2
django-filter==0.12.0
